fix: apply sinc2 correction to volumes from torch-fourier-slice
