+++ src/leopard_em/utils/fourier_slice.py
@@ -3,19 +3,36 @@
 import roma
 import torch
 from torch_fourier_slice import extract_central_slices_rfft_3d
-from torch_grid_utils import fftfreq_grid
+from torch_fourier_slice.volume_utils import separable_sinc2_correction
 
 
-def _sinc2(shape: tuple[int, ...], rfft: bool, fftshift: bool) -> torch.Tensor:
-    """Helper function for creating a sinc^2 filter."""
-    grid = fftfreq_grid(
-        image_shape=shape,
-        rfft=rfft,
-        fftshift=fftshift,
-        norm=True,
-    )
+def volume_to_rfft_fourier_slice(volume: torch.Tensor) -> torch.Tensor:
+    """Prepares a 3D volume for Fourier slice extraction.
 
-    return torch.sinc(grid) ** 2
+    Parameters
+    ----------
+    volume : torch.Tensor
+        The input volume.
+
+    Returns
+    -------
+    torch.Tensor
+        The prepared volume in Fourier space ready for slice extraction.
+    """
+    assert volume.dim() == 3, "Volume must be 3D"
+
+    sinc2 = separable_sinc2_correction(volume.shape, device=volume.device)
+    volume = volume / sinc2
+
+    # NOTE: There is an extra FFTshift step before the RFFT since, for some reason,
+    # omitting this step will cause a 180 degree phase shift on odd (i, j, k)
+    # structure factors in the Fourier domain. This just requires an extra
+    # IFFTshift after converting a slice back to real-space (handled already).
+    volume = torch.fft.fftshift(volume, dim=(0, 1, 2))  # pylint: disable=E1102
+    volume_rfft = torch.fft.rfftn(volume, dim=(0, 1, 2))  # pylint: disable=E1102
+    volume_rfft = torch.fft.fftshift(volume_rfft, dim=(0, 1))  # pylint: disable=E1102
+
+    return volume_rfft
 
 
 def _rfft_slices_to_real_projections(
@@ -71,9 +88,7 @@ def get_rfft_slices_from_volume(
         The Fourier slices of the volume.
 
     """
-    volume_rfft = torch.fft.fftshift(volume, dim=(-3, -2, -1))  # pylint: disable=not-callable
-    volume_rfft = torch.fft.fftn(volume_rfft, dim=(-3, -2, -1))  # pylint: disable=not-callable
-    volume_rfft = torch.fft.fftshift(volume_rfft, dim=(-3, -2))  # pylint: disable=not-callable
+    volume_rfft = volume_to_rfft_fourier_slice(volume)
 
     # Use roma to keep angles on same device
     rot_matrix = roma.euler_to_rotmat("ZYZ", (phi, theta, psi), degrees=degrees)
